Move helper to utils file
